Work Page Added
Added a work page so that we could track what work we’ve done as that’s
part of the project that we haven’t done yet. Remember though that we
didn’t have a group for the first week this project was assigned so no
one did anything till week 2.